Added end to end testcase
